refactor(presets): split oversized modules — core.js & manager.js
core.js 902 -> 768 lines:
- snapshot.js (new, 121): snapshotOf / stripSystemKeys / hasKeyChanged /
  captureCurrentSettings / applySnapshot — pure functions, no module state.
- builtin.js (new, 52): builtin preset loading (list + per-file parse), own cache.
- core.js keeps the settings-stream handler, CRUD, write-back, init; re-exports
  getBuiltinPresets/getBuiltinSettings/captureCurrentSettings/applySnapshot so
  existing importers (manager.js) are unchanged.

manager.js 1015 -> 706 lines:
- form.js (new, 246): settings form rendering (renderForm/buildSettingRow/
  buildControl/syncFormControls/fillFormFromDefaults/selectAllCheckboxes/
  invertCheckboxes/collectCheckedSnapshot) via createForm(api) — manager.js
  injects entries/formValues/formIncluded/formEl/escapeHtml, stays the owner
  of page state.

Verified: sim suite 22 scenarios PASS; all modules node --check; import
resolution check passes; Playwright local-server UI check renders 36 setting
rows / 13 list items / guide button / 16x16 checkboxes with zero JS errors.
